+++ directx12-lib/src/Mesh.h
@@ -8,31 +8,31 @@
 
 struct Vertex
 {
-    DirectX::XMFLOAT3 Position;
-    DirectX::XMFLOAT3 Normal;
-    DirectX::XMFLOAT2 UV;
-    DirectX::XMFLOAT3 Tangent;
-    DirectX::XMFLOAT4 Color;
+	DirectX::XMFLOAT3 Position;
+	DirectX::XMFLOAT3 Normal;
+	DirectX::XMFLOAT2 UV;
+	DirectX::XMFLOAT3 Tangent;
+	DirectX::XMFLOAT4 Color;
 };
 
 struct Material
 {
-    std::string diffuse_map_name = "";
-    D3D12_GPU_DESCRIPTOR_HANDLE handle_gpu = {};
-    DirectX::XMFLOAT4 diffuse_color = { 1,1,1,1 };
+	std::string diffuse_map_name = "";
+	D3D12_GPU_DESCRIPTOR_HANDLE handle_gpu = {};
+	DirectX::XMFLOAT4 diffuse_color = { 1,1,1,1 };
 };
 
 /*
 struct Bone {
-    std::string name = "";
-    aiMatrix4x4 transform = {};
+	std::string name = "";
+	aiMatrix4x4 transform = {};
 };
 */
 
 struct Mesh
 {
-    std::vector<Vertex> vertices;
-    std::vector<uint32_t> indices;
-    Material material;
-    //std::vector<Bone> bones;
+	std::vector<Vertex> vertices;
+	std::vector<unsigned short> indices;
+	Material material;
+	//std::vector<Bone> bones;
 };
